requirements.txt
# ── Web framework ─────────────────────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6

# ── Streamlit UI ───────────────────────────────────────────────────────────────
streamlit==1.38.0

# ── Database ───────────────────────────────────────────────────────────────────
sqlalchemy==2.0.36

# ── Data validation ────────────────────────────────────────────────────────────
pydantic==2.9.2
pydantic-settings==2.5.2

# ── LLM provider (Anthropic Claude) ───────────────────────────────────────────
anthropic==0.34.2

# ── HTTP client (used internally by anthropic SDK) ────────────────────────────
httpx==0.27.2

# ── CSV / data handling ────────────────────────────────────────────────────────
pandas==2.2.3

# ── Env / config ──────────────────────────────────────────────────────────────
python-dotenv==1.0.1

# ── Multipart file upload support for FastAPI ─────────────────────────────────
python-multipart==0.0.9

# ── Testing ────────────────────────────────────────────────────────────────────
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2          # also used as async test client for FastAPI

# ── Utilities ─────────────────────────────────────────────────────────────────
tenacity==9.0.0        # retry logic with backoff
python-dateutil==2.9.0
